+++ src/test/java/com/group16/uno/service/DailyScoreServiceTest.java
@@ -12,6 +12,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.util.Optional;
 import java.util.List;
@@ -41,7 +42,7 @@ class DailyScoreServiceTest {
         user = new User( "johndoe", "john@example.com", "hashedPassword");
         user.setId("userId");
         createdAt = Date.valueOf("2025-04-04");
-        dailyScore = new DailyScore(user, createdAt, 100);
+        dailyScore = new DailyScore(user, createdAt, BigDecimal.valueOf(100));
     }
 
     @Test
@@ -51,7 +52,7 @@ class DailyScoreServiceTest {
         when(dailyScoreRepository.save(any(DailyScore.class))).thenReturn(dailyScore);
 
         // Act
-        DailyScore result = dailyScoreService.createDailyScore("userId", createdAt, 100);
+        DailyScore result = dailyScoreService.createDailyScore("userId", createdAt, BigDecimal.valueOf(100));
 
         // Assert
         assertNotNull(result);
@@ -68,7 +69,7 @@ class DailyScoreServiceTest {
 
         // Act & Assert
         RuntimeException exception = assertThrows(RuntimeException.class, () -> {
-            dailyScoreService.createDailyScore("invalidUserId", createdAt, 100);
+            dailyScoreService.createDailyScore("invalidUserId", createdAt, BigDecimal.valueOf(100));
         });
         assertEquals("User not found with id: invalidUserId", exception.getMessage());
         verify(userRepository).findById("invalidUserId");
@@ -92,19 +93,19 @@ class DailyScoreServiceTest {
     @Test
     void updateDailyScore_shouldUpdateScore() {
         // Arrange
-        doNothing().when(dailyScoreRepository).updateDailyScore("userId", createdAt, 200);
+        doNothing().when(dailyScoreRepository).updateDailyScore("userId", createdAt, BigDecimal.valueOf(200));
 
         // Act
-        dailyScoreService.updateDailyScore("userId", createdAt, 200);
+        dailyScoreService.updateDailyScore("userId", createdAt, BigDecimal.valueOf(200));
 
         // Assert
-        verify(dailyScoreRepository).updateDailyScore("userId", createdAt, 200);
+        verify(dailyScoreRepository).updateDailyScore("userId", createdAt, BigDecimal.valueOf(200));
     }
 
     @Test
     void getWeeklyLeaderBoard_shouldReturnLeaderBoard() {
         // Arrange
-        List<LeaderBoardDTO> leaderBoard = List.of(new LeaderBoardDTO("userId", 100));
+        List<LeaderBoardDTO> leaderBoard = List.of(new LeaderBoardDTO("userId", BigDecimal.valueOf(100)));
         when(dailyScoreRepository.getWeeklyLeaderBoard()).thenReturn(leaderBoard);
 
         // Act
@@ -120,7 +121,7 @@ class DailyScoreServiceTest {
     @Test
     void getMonthlyLeaderBoard_shouldReturnLeaderBoard() {
         // Arrange
-        List<LeaderBoardDTO> leaderBoard = List.of(new LeaderBoardDTO("userId", 100));
+        List<LeaderBoardDTO> leaderBoard = List.of(new LeaderBoardDTO("userId", BigDecimal.valueOf(100)));
         when(dailyScoreRepository.getMonthlyLeaderBoard()).thenReturn(leaderBoard);
 
         // Act
@@ -136,7 +137,7 @@ class DailyScoreServiceTest {
     @Test
     void getAllTimeLeaderBoard_shouldReturnLeaderBoard() {
         // Arrange
-        List<LeaderBoardDTO> leaderBoard = List.of(new LeaderBoardDTO("userId", 100));
+        List<LeaderBoardDTO> leaderBoard = List.of(new LeaderBoardDTO("userId", BigDecimal.valueOf(100)));
         when(dailyScoreRepository.getAllTimeLeaderBoard()).thenReturn(leaderBoard);
 
         // Act
